refactor react-native packages
