+++ packages/plugin-react-native-client-sync/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../../tsconfig.json",
   "include": ["."],
   "compilerOptions": {
-    "lib": ["es2015"],
+    "lib": ["es2015", "esnext"],
     "types": ["jest", "react-native"],
   }
 }
