+++ packages/plugin-react-native-orientation-breadcrumbs/package.json
@@ -17,7 +17,8 @@
   "author": "Bugsnag",
   "license": "MIT",
   "devDependencies": {
-    "@bugsnag/core": "^8.6.0"
+    "@bugsnag/core": "^8.6.0",
+    "@types/react-native": "0.67.8"
   },
   "peerDependencies": {
     "@bugsnag/core": "^8.0.0"
