+++ packages/react-native/package.json
@@ -48,6 +48,9 @@
   "license": "MIT",
   "devDependencies": {
     "@babel/cli": "^7.0.0",
+    "@types/jest": "^26",
+    "@types/react-native": "0.67.8",
+    "react-native": "^0.63.4",
     "tslint": "^5.12.1",
     "typescript": "^3.3.3"
   },
@@ -63,7 +66,8 @@
     "@bugsnag/plugin-react-native-hermes": "^8.6.0",
     "@bugsnag/plugin-react-native-session": "^8.6.0",
     "@bugsnag/plugin-react-native-unhandled-rejection": "^8.6.0",
-    "iserror": "^0.0.2"
+    "iserror": "^0.0.2",
+    "react-native": "^0.63.4"
   },
   "scripts": {
     "prepare": "bash prepare-android-vendor.sh",
