+++ packages/react-native/src/NativeBugsnag.ts
@@ -46,14 +46,14 @@ export interface Spec extends TurboModule {
   updateGroupingDiscriminator(groupingDiscriminator: string | undefined | null): void
 }
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+ 
 export type ReactNativeConfiguration = {
   reactNativeVersion?: string
   engine?: string
   notifierVersion: string
 }
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+ 
 export type NativeConfiguration = {
   apiKey: string
   autoDetectErrors?: boolean
@@ -75,7 +75,7 @@ export type NativeConfiguration = {
   endpoints: NativeEndpointConfig
 }
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+ 
 export type NativeBreadcrumb = {
   timestamp: string
   message: string
@@ -83,13 +83,13 @@ export type NativeBreadcrumb = {
   metadata?: UnsafeObject
 }
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+ 
 export type NativeEndpointConfig = {
   notify: string
   sessions: string
 }
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+ 
 export type NativeEnabledErrorTypes = {
   anrs?: boolean
   ndkCrashes?: boolean
@@ -97,7 +97,7 @@ export type NativeEnabledErrorTypes = {
   unhandledRejections?: boolean
 }
 
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+ 
 export type NativeFeatureFlag = {
   name: string
   variant?: string
