+++ packages/react-native/src/notifier.js
@@ -19,10 +19,7 @@ const url = 'https://github.com/bugsnag/bugsnag-js'
 
 const React = require('react')
 
-const Client = require('@bugsnag/core/client')
-const Event = require('@bugsnag/core/event')
-const Session = require('@bugsnag/core/session')
-const Breadcrumb = require('@bugsnag/core/breadcrumb')
+const { Breadcrumb, Client, Event, Session } = require('@bugsnag/core')
 
 Event.__type = 'reactnativejs'
