+++ packages/react-native/tsconfig.json
@@ -1,8 +1,16 @@
 {
-  "extends": "../../tsconfig.json",
-  "include": ["."],
   "compilerOptions": {
+    "allowJs": true,
+    "jsx": "preserve",
+    "noEmit": true,
+    "strict": true,
+    "esModuleInterop": true,
+    "target": "es5",
     "lib": ["es2015"],
     "types": ["jest", "react-native"],
-  }
+  },
+  "include": ["."],
+  "exclude": [
+    "src/NativeBugsnag.ts"
+  ],
 }
